+работа с изображениями в посте (добавление, удаление) +относительный контроль POST,GET,COOKIE,SERVER -bootstrap

+++ models/db.php
@@ -3,12 +3,14 @@
 class MySQLdata {
 
     protected $mysqli;
+    public $error;
 
     function __construct() {
         $this->DB_HOST = DB_HOST;
         $this->DB_LOGIN = DB_LOGIN;
         $this->DB_PASSWORD = DB_PASSWORD;
         $this->DB_NAME = DB_NAME;
+        $this->error = false;
         $this->connect();
     }
 
@@ -18,52 +20,114 @@ class MySQLdata {
 
     function connect() {
         /* проверка соединения */
-        $this->mysqli = new mysqli($this->DB_HOST, $this->DB_LOGIN, $this->DB_PASSWORD, $this->DB_NAME);
-        if (mysqli_connect_errno()) {
-            printf("Не удалось подключиться: %s\n", mysqli_connect_error());
+        mysqli_report(MYSQLI_REPORT_STRICT);
+        try {
+            $this->mysqli = new mysqli($this->DB_HOST, $this->DB_LOGIN, $this->DB_PASSWORD, $this->DB_NAME);
+        } catch (Exception $e) {
+            $this->error = "Can't connect to database: " . mysqli_connect_error();
+            echo $this->error;
             exit();
         }
-        /* изменение набора символов на utf8 */
         if (!$this->mysqli->set_charset("utf8")) {
-            printf("Ошибка при загрузке набора символов utf8: %s\n", $this->mysqli->error);
+            $this->error = "Can't change charset utf8: " . $this->mysqli->error;
+            echo $this->error;
+            exit();
         }
     }
 
     function disconnect() {
-        $this->mysqli->close();
+        if (!$this->error) {
+            $this->mysqli->close();
+        }
     }
 
-    function addUserAndGetSessionHash($username, $password) {
-
-        $sessionhash = getHash(time() + rand());
-        $passhash = getHash($password);
-        $sqlquery = "INSERT INTO users (login, passhash, sesshash) VALUES ('$username', '$passhash', '$sessionhash')";
-        $sqlresult = $this->mysqli->query($sqlquery);
-        return $sessionhash;
-    }
-
-    function getNewSessionHashAndUserID($username, $password) {
-        $sessionhash = getHash(time() + rand());
+    function addUserGetSessionHashAndID($username, $password) {
+        $username = $this->mysqli->real_escape_string($username);
+        $password = $this->mysqli->real_escape_string($password);
         $sqlquery = "SELECT * FROM users WHERE LOWER(login) = LOWER('$username')";
-        if ($sqlresult = $this->mysqli->query($sqlquery)) {
-            $tablerow = $sqlresult->fetch_assoc();
-        }
-
-        if (getHash($password) !== $tablerow['passhash']) {
-            $sessionhash = FALSE;
+        $sqlresult = $this->mysqli->query($sqlquery);
+        if ($sqlresult->num_rows > 0) {
+            $credentials['error'] = 'Choose other login!';
         } else {
-            $sqlquery = "UPDATE users SET `sesshash` = '$sessionhash' WHERE LOWER(login) = LOWER('$username')";
+            $newsessionhash = getHash(time() + rand());
+            $passhash = getHash($password);
+            $sqlquery = "INSERT INTO users (login, passhash, sesshash) VALUES ('$username', '$passhash', '$newsessionhash')";
             $sqlresult = $this->mysqli->query($sqlquery);
+            $credentials['hashsess'] = $newsessionhash;
+            $credentials['userid'] = $this->mysqli->insert_id;
+            $credentials['error'] = false;
         }
-        $sessandid['hash'] = $sessionhash;
-        $sessandid['userid'] = $tablerow['id'];
-        return $sessandid;
+        return $credentials;
     }
 
-    function isUserAuthent($userid, $sesshash) {
+    function getUserCredentials($username, $password) {
+        $username = $this->mysqli->real_escape_string($username);
+        $password = $this->mysqli->real_escape_string($password);
+        $sqlquery = "SELECT * FROM users WHERE LOWER(login) = LOWER('$username')";
+        $sqlresult = $this->mysqli->query($sqlquery);
+        if ($sqlresult->num_rows > 0) {
+            $tablerow = $sqlresult->fetch_assoc();
+            if (getHash($password) == $tablerow['passhash']) {
+                $newsessionhash = getHash(time() + rand());
+                $sqlquery = "UPDATE users SET `sesshash` = '$newsessionhash' WHERE LOWER(login) = LOWER('$username')";
+                $sqlresult = $this->mysqli->query($sqlquery);
+                $credentials['hashsess'] = $newsessionhash;
+                $credentials['userid'] = $tablerow['id'];
+                $credentials['email'] = $tablerow['email'];
+                $credentials['fullname'] = $tablerow['fullname'];
+                $credentials['username'] = $tablerow['login'];
+                $credentials['error'] = false;
+            } else {
+                $credentials['error'] = 'Wrong password!';
+            }
+        } else {
+            $credentials['error'] = 'No such login!';
+        }
+        return $credentials;
+    }
+
+    function getUserInfo($userid, $hashsess) {
+        $userid = $this->mysqli->real_escape_string($userid);
+        $hashsess = $this->mysqli->real_escape_string($hashsess);
+        $sqlquery = "SELECT * FROM users WHERE id=$userid and sesshash='$hashsess'";
+        $sqlresult = $this->mysqli->query($sqlquery);
+        if ($sqlresult->num_rows > 0) {
+            $tablerow = $sqlresult->fetch_assoc();
+            $credentials['email'] = $tablerow['email'];
+            $credentials['fullname'] = $tablerow['fullname'];
+            $credentials['username'] = $tablerow['login'];
+            $credentials['role'] = $tablerow['role'];
+            $credentials['error'] = false;
+        } else {
+            $credentials['error'] = 'No such user!';
+        }
+        return $credentials;
+    }
+
+    function editUserInfo($userid, $hashsess, $fullname, $email) {
+        $userid = $this->mysqli->real_escape_string($userid);
+        $hashsess = $this->mysqli->real_escape_string($hashsess);
+        $fullname = $this->mysqli->real_escape_string($fullname);
+        $email = $this->mysqli->real_escape_string($email);
+        $sqlquery = "UPDATE `users` SET `fullname` = '$fullname',`email` = '$email' where id=$userid and sesshash='$hashsess'";
+        $sqlresult = $this->mysqli->query($sqlquery);
+    }
+
+    function destroyUserSession($userid, $hashsess) {
+        $userid = $this->mysqli->real_escape_string($userid);
+        $hashsess = $this->mysqli->real_escape_string($hashsess);
+        $randomhash = getHash(time() + rand());
+        $sqlquery = "UPDATE `users` SET `sesshash` = '$randomhash' where id=$userid and sesshash='$hashsess'";
+        $sqlresult = $this->mysqli->query($sqlquery);
+    }
+
+    function isUserAuthent($userid, $hashsess) {
+        $userid = $this->mysqli->real_escape_string($userid);
+        $hashsess = $this->mysqli->real_escape_string($hashsess);
         $role = false;
-        $sqlquery = "SELECT * FROM users WHERE id = '$userid' and sesshash = '$sesshash'";
-        if ($sqlresult = $this->mysqli->query($sqlquery)) {
+        $sqlquery = "SELECT * FROM users WHERE id = '$userid' and sesshash = '$hashsess'";
+        $sqlresult = $this->mysqli->query($sqlquery);
+        if ($sqlresult->num_rows > 0) {
             $tablerow = $sqlresult->fetch_assoc();
             $role = $tablerow['role'];
         }
@@ -71,9 +135,11 @@ class MySQLdata {
     }
 
     function getUserRole($userid) {
+        $userid = $this->mysqli->real_escape_string($userid);
         $role = false;
         $sqlquery = "SELECT * FROM users WHERE id = '$userid'";
-        if ($sqlresult = $this->mysqli->query($sqlquery)) {
+        $sqlresult = $this->mysqli->query($sqlquery);
+        if ($sqlresult->num_rows > 0) {
             $tablerow = $sqlresult->fetch_assoc();
             $role = $tablerow['role'];
         }
@@ -81,7 +147,10 @@ class MySQLdata {
     }
 
     function getPrePosts($offset, $rowcount, $previewlength) {
-        $sqlquery = "SELECT posts.id,login,CAST(text as char($previewlength))as text,title,date FROM `posts` inner join `users` on posts.userid=users.id LIMIT $offset,$rowcount";
+        $offset = $this->mysqli->real_escape_string($offset);
+        $rowcount = $this->mysqli->real_escape_string($rowcount);
+        $previewlength = $this->mysqli->real_escape_string($previewlength);
+        $sqlquery = "SELECT posts.id,login,fullname,CAST(text as char($previewlength))as text,title,date FROM `posts` inner join `users` on posts.userid=users.id LIMIT $offset,$rowcount";
         if ($sqlresult = $this->mysqli->query($sqlquery)) {
             while ($row = $sqlresult->fetch_assoc()) {
                 $tablerow[] = $row;
@@ -91,7 +160,11 @@ class MySQLdata {
     }
 
     function getPrePostsByTag($offset, $rowcount, $previewlength, $tag) {
-        $sqlquery = "SELECT posts.id,login,CAST(text as char($previewlength))as text,title,date from posts inner join tagbp inner join tags inner join `users` on posts.userid=users.id and posts.id=tagbp.postid and tagbp.tagid=tags.id where tags.id=$tag LIMIT $offset,$rowcount";
+        $offset = $this->mysqli->real_escape_string($offset);
+        $rowcount = $this->mysqli->real_escape_string($rowcount);
+        $previewlength = $this->mysqli->real_escape_string($previewlength);
+        $tag = $this->mysqli->real_escape_string($tag);
+        $sqlquery = "SELECT posts.id,login,fullname,CAST(text as char($previewlength))as text,title,date from posts inner join tagbp inner join tags inner join `users` on posts.userid=users.id and posts.id=tagbp.postid and tagbp.tagid=tags.id where tags.id=$tag LIMIT $offset,$rowcount";
         if ($sqlresult = $this->mysqli->query($sqlquery)) {
             while ($row = $sqlresult->fetch_assoc()) {
                 $tablerow[] = $row;
@@ -101,6 +174,7 @@ class MySQLdata {
     }
 
     function getPost($postid) {
+        $postid = $this->mysqli->real_escape_string($postid);
         $sqlquery = "SELECT posts.id,login,text,title,date FROM `posts` inner join `users` on posts.userid=users.id WHERE posts.id=$postid";
         if ($sqlresult = $this->mysqli->query($sqlquery)) {
             $row = $sqlresult->fetch_assoc();
@@ -109,6 +183,7 @@ class MySQLdata {
     }
 
     function getPostTags($postid) {
+        $postid = $this->mysqli->real_escape_string($postid);
         $res = False;
         $sqlquery = "select tags.id,tags.tag from posts inner join tags inner join tagbp on tags.id=tagbp.tagid and posts.id=tagbp.postid where posts.id=$postid";
         $sqlresult = $this->mysqli->query($sqlquery);
@@ -122,6 +197,7 @@ class MySQLdata {
     }
 
     function newTag($tag) {
+        $tag = $this->mysqli->real_escape_string($tag);
         $sqlquery = "select * from tags where tag ='$tag'";
         if ($sqlresult = $this->mysqli->query($sqlquery)) {
             $row = $sqlresult->fetch_assoc();
@@ -136,6 +212,7 @@ class MySQLdata {
     }
 
     function getTagsByPart($part) {
+        $part = $this->mysqli->real_escape_string($part);
         $res = False;
         $sqlquery = "select * from tags where tag like '%$part%' limit 10";
         $sqlresult = $this->mysqli->query($sqlquery);
@@ -149,10 +226,15 @@ class MySQLdata {
     }
 
     function addTagsToPost($postid, $tags) {
+        $postid = $this->mysqli->real_escape_string($postid);
         foreach ($tags as $id => $tag) {
+            $id = $this->mysqli->real_escape_string($id);
+            $tag = $this->mysqli->real_escape_string($tag);
+            $tags[$id] = $this->mysqli->real_escape_string($tags[$id]);
+
             if (preg_match("/new\d+$/", $id)) {
                 $newtagid = $this->newTag($tag);
-                $tags[$newtagid] = $arr[$id];
+                $tags[$newtagid] = $tags[$id];
                 unset($tags[$id]);
             }
         }
@@ -160,6 +242,7 @@ class MySQLdata {
         $sqlresult = $this->mysqli->query($sqlquery);
         $sqlquery = "insert into tagbp (postid, tagid) VALUES ";
         foreach ($tags as $id => $tag) {
+            $id = $this->mysqli->real_escape_string($id);
             $sqlquery = $sqlquery . "(" . $postid . "," . $id . "),";
         }
         $sqlquery = mb_substr($sqlquery, 0, -1);
@@ -167,6 +250,7 @@ class MySQLdata {
     }
 
     function getPostFiles($postid) {
+        $postid = $this->mysqli->real_escape_string($postid);
         $res = null;
         $sqlquery = "select * from images where postid='$postid'";
         $sqlresult = $this->mysqli->query($sqlquery);
@@ -180,10 +264,10 @@ class MySQLdata {
     }
 
     function addFileToPost($postid, $files) {
-        //$sqlquery = "delete from images where postid = '$postid'";
-        //$sqlresult = $this->mysqli->query($sqlquery);
+        $postid = $this->mysqli->real_escape_string($postid);
         $sqlquery = "insert into images (postid, filename) VALUES ";
         foreach ($files as $id => $fn) {
+            $fn = $this->mysqli->real_escape_string($fn);
             $sqlquery = $sqlquery . "(" . $postid . ",'" . $fn . "'),";
         }
         $sqlquery = mb_substr($sqlquery, 0, -1);
@@ -191,26 +275,55 @@ class MySQLdata {
     }
 
     function delFileFromPost($postid, $userid, $hashsess, $file) {
-        $file = basename($file);
+        $postid = $this->mysqli->real_escape_string($postid);
+        $userid = $this->mysqli->real_escape_string($userid);
+        $hashsess = $this->mysqli->real_escape_string($hashsess);
+        $file = $this->mysqli->real_escape_string($file);
+        $error = false;
         if ($role = $this->isUserAuthent($userid, $hashsess)) {
-            unlink(ROOT . uploaddir . $file);
-            $sqlquery = "delete from images where postid = '$postid' and filename='$file'";
-            $sqlresult = $this->mysqli->query($sqlquery);
+            $file = basename($file);
+            $sqlquery = "select * from posts where id = $postid";
+            if ($sqlresult = $this->mysqli->query($sqlquery)) {
+                $row = $sqlresult->fetch_assoc();
+                $postsuserid = $row['userid'];
+            }
+            if ($userid == $postsuserid or $role == ADMIN_ROLE) {
+                unlink(ROOT . uploaddir . $file);
+                $sqlquery = "delete from images where postid = '$postid' and filename='$file'";
+                $sqlresult = $this->mysqli->query($sqlquery);
+            } else {
+                $error = "Вы не можете редактировать эту статью.";
+            }
         }
+        return $error;
     }
 
     function updatePost($postid, $title, $text, $userid, $hashsess, $tags, $files) {
+        $postid = $this->mysqli->real_escape_string($postid);
+        $userid = $this->mysqli->real_escape_string($userid);
+        $hashsess = $this->mysqli->real_escape_string($hashsess);
+        $title = $this->mysqli->real_escape_string($title);
+        $text = $this->mysqli->real_escape_string($text);
+        $error = false;
         if ($role = $this->isUserAuthent($userid, $hashsess)) {
             $now = time();
             $sqlquery = "UPDATE `posts` SET `title` = '$title', `text` = '$text', `date` = '$now' where (posts.id=$postid and (posts.userid='$userid' or $role=1))";
-            $sqlresult = $this->mysqli->query($sqlquery);
-            $this->addTagsToPost($postid, $tags);
-            $this->addFileToPost($postid, $files);
+            if ($sqlresult = $this->mysqli->query($sqlquery)) {
+                $this->addTagsToPost($postid, $tags);
+                $this->addFileToPost($postid, $files);
+            } else {
+                $error = 'Вы не можете редактировать эту статью.';
+            }
         }
+        return $error;
 //   большой запрос, сразу с авторизацией (как вариант)                $sqlquery = "UPDATE `posts` as tempposts inner join users as userstemp1 on userstemp1.id=tempposts.userid inner join users as userstemp2 on userstemp2.id=$userid SET `title` = '$title', `text` = '$text', `date` = '$now' where ((tempposts.userid=$userid and userstemp1.sesshash='$hashsess') or (userstemp2.role=1 and userstemp2.sesshash='$hashsess')) and tempposts.id=$postid";
     }
 
     function newPost($title, $text, $userid, $hashsess, $tags, $files) {
+        $userid = $this->mysqli->real_escape_string($userid);
+        $hashsess = $this->mysqli->real_escape_string($hashsess);
+        $title = $this->mysqli->real_escape_string($title);
+        $text = $this->mysqli->real_escape_string($text);
         if ($role = $this->isUserAuthent($userid, $hashsess)) {
             $now = time();
             $sqlquery = "insert into posts (userid,text,title,date) VALUES ('$userid','$text','$title', '$now')";
@@ -224,11 +337,15 @@ class MySQLdata {
     }
 
     function erasePost($postid, $userid, $hashsess) {
+        $postid = $this->mysqli->real_escape_string($postid);
+        $userid = $this->mysqli->real_escape_string($userid);
+        $hashsess = $this->mysqli->real_escape_string($hashsess);
         $sqlquery = "DELETE from tempposts using `posts` as tempposts inner join users as userstemp1 on userstemp1.id=tempposts.userid inner join users as userstemp2 on userstemp2.id=$userid where ((tempposts.userid=$userid and userstemp1.sesshash='$hashsess') or (userstemp2.role=1 and userstemp2.sesshash='$hashsess')) and tempposts.id=$postid";
         $sqlresult = $this->mysqli->query($sqlquery);
     }
 
     function getPostAmount($tag) {
+        $tag = $this->mysqli->real_escape_string($tag);
         if ($tag) {
             $sqlquery = "SELECT COUNT(*) from posts inner join tagbp inner join tags inner join `users` on posts.userid=users.id and posts.id=tagbp.postid and tagbp.tagid=tags.id where tags.id='$tag'";
         } else {
@@ -242,8 +359,14 @@ class MySQLdata {
     }
 
     function recordLog($userid, $uniq, $server) {
-        $now = time();
+        $userid = $this->mysqli->real_escape_string($userid);
+        $uniq = $this->mysqli->real_escape_string($uniq);
+        $server['HTTP_USER_AGENT'] = $this->mysqli->real_escape_string($server['HTTP_USER_AGENT']);
+        @$server["HTTP_REFERER"] = $this->mysqli->real_escape_string($server["HTTP_REFERER"]);
+        $server["REMOTE_ADDR"] = $this->mysqli->real_escape_string($server["REMOTE_ADDR"]);
 
+
+        $now = time();
         if (!$user_agent = @$server['HTTP_USER_AGENT']) {
             $user_agent = "";
         }
@@ -267,6 +390,7 @@ class MySQLdata {
     }
 
     function getCommentByPost($postid) {
+        $postid = $this->mysqli->real_escape_string($postid);
         $res = False;
         $sqlquery = "select * from comments inner join users on comments.userid=users.id where comments.postid=$postid";
         $sqlresult = $this->mysqli->query($sqlquery);
@@ -280,6 +404,10 @@ class MySQLdata {
     }
 
     function addComment($postid, $userid, $hashsess, $text) {
+        $postid = $this->mysqli->real_escape_string($postid);
+        $userid = $this->mysqli->real_escape_string($userid);
+        $hashsess = $this->mysqli->real_escape_string($hashsess);
+        $text = $this->mysqli->real_escape_string($text);
         if ($role = $this->isUserAuthent($userid, $hashsess)) {
             $now = time();
             $sqlquery = "insert into comments (postid,userid,text,date) VALUES ('$postid','$userid','$text','$now')";
